trojsten/login/serializers.py
from rest_framework import serializers

from trojsten.people.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ("id", "username", "email")
